Modification de la classe Exchange et rectification des classes models

+++ src/VRP/model/HotelNode.java
@@ -1,7 +1,7 @@
 package VRP.model;
 
 public class HotelNode extends Node{
-    public HotelNode(int id, double score) {
+    public HotelNode(int id, int score) {
         super(id, score);
     }
